+++ include/pmDomain.h
@@ -30,14 +30,14 @@
 class pmDomain {
 	pmTensor minimum;
 	pmTensor maximum;
-	float cell_size;
+	double cell_size;
 	pmTensor boundary;
 public:
 	pmDomain() {}
-	pmDomain(pmTensor const& dmin, pmTensor const& dmax, float const& csize, pmTensor const& bnd);
+	pmDomain(pmTensor const& dmin, pmTensor const& dmax, double const& csize, pmTensor const& bnd);
 	pmTensor get_minimum() const;
 	pmTensor get_maximum() const;
-	float get_cell_size() const;
+	double get_cell_size() const;
 	size_t get_num_cells() const;
 	size_t get_dimensions() const;
 	pmTensor get_physical_minimum() const;
@@ -46,7 +46,7 @@ public:
 	pmTensor get_boundary() const;
 	void set_minimum(pmTensor const& mn);
 	void set_maximum(pmTensor const& mx);
-	void set_cell_size(float const& csize);
+	void set_cell_size(double const& csize);
 	void set_boundary(pmTensor const& bnd);
 	void print() const;
 };
